requirements.txt
Django==4.2.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7